Development Stage Completed
Program now consists of the following important files:
GUI.py - Which contains all of the GUI and a small section of main code.
Functions.py - Which contains the functions previously stored in FunctionsCLI.py
fileListSetName.txt - Which contains the name of each set and the file the set is stored in.

NB: The ClassListImplementation.py File was used when the program went from using nested lists to a list class system.